[Feat] 프로필 수정, 전체 레이아웃 수정
