+++ src/components/common/Post/PostStyle.jsx
@@ -2,27 +2,33 @@ import styled from "styled-components";
 
 export const PostArticle = styled.article`
   display: flex;
+  flex-direction: column;
+`;
+export const PostFlexWrap = styled.div`
+  display: flex;
 `;
 
-export const Wrap = styled.div`
-  width: 304px;
-  margin-left: 12px;
-  margin-top: 4px;
+export const ProfileButton = styled.button`
+  display: flex;
 `;
 
 export const PostProfileImg = styled.img`
   width: 42px;
   height: 42px;
+  border-radius: 50%;
+  border: 1px solid #dbdbdb;
 `;
 
 export const PostNameWrap = styled.div`
   display: flex;
   flex-direction: column;
+  gap: 5px;
+  margin-left: 13.5px;
 `;
 
 export const UserSpan = styled.span`
   font-size: 14px;
-  font-weight: 500;
+  font-weight: bold;
 `;
 
 export const AccountSpan = styled.span`
@@ -32,14 +38,70 @@ export const AccountSpan = styled.span`
   color: rgba(118, 118, 118, 1);
 `;
 
-export const PostFlexWrap = styled.div`
-  display: flex;
-  justify-content: space-between;
+export const Time = styled.p`
+  margin-left: 70px;
+  padding-top: 15px;
+  font-weight: bold;
+  margin-right: 10px;
+  & > span {
+    display: inline-block;
+    margin-left: 5px;
+    font-weight: 400;
+    width: 20px;
+    text-align: center;
+  }
+`;
+
+export const DotButton = styled.button`
+  width: 18px;
+  height: 18px;
 `;
 
 export const DotImg = styled.img`
   width: 18px;
   height: 18px;
+`;
+
+export const Wrap = styled.div`
+  width: 358px;
+  padding-left: 30px;
+`;
+
+export const FeedButton = styled.button``;
+
+export const HealthWrap = styled.ul`
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+  margin-top: 22px;
+  margin-bottom: 11px;
+  padding-bottom: 11px;
+  font-size: 14px;
+  border-bottom: 1px solid #dbdbdb;
+`;
+
+export const HealthList = styled.li`
+  font-weight: bold;
+  & > span {
+    display: inline-block;
+    text-align: center;
+    font-weight: 400;
+  }
+  & > span:first-child {
+    width: 112px;
+    margin-right: 25px;
+  }
+  & > span:nth-child(2) {
+    width: 40px;
+  }
+  & > span:nth-child(3) {
+    width: 30px;
+    margin-left: 10px;
+  }
+  & > span:last-child {
+    margin-left: 10px;
+    width: 30px;
+  }
 `;
 
 export const PostContent = styled.div`
@@ -48,7 +110,7 @@ export const PostContent = styled.div`
 `;
 
 export const PostUploadImg = styled.img`
-  width: 304px;
+  width: 328px;
   border-radius: 10px;
 `;
 
@@ -61,7 +123,6 @@ export const HeartImg = styled.img`
 export const MessageImg = styled.img`
   width: 20px;
   height: 20px;
-  margin-left: 16px;
   margin-right: 6px;
 `;
 
@@ -79,3 +140,9 @@ export const ButtonWrap = styled.div`
 export const HeartSpan = styled.span``;
 
 export const MessageSpan = styled.span``;
+
+export const HeartButton = styled.button``;
+
+export const MessageButton = styled.button`
+  margin-left: 16px;
+`;
