+++ src/pages/Home/Search/Search.jsx
@@ -1,15 +1,14 @@
-import React from 'react'
-import SearchNav from '../../../components/Header/SearchHeader'
-import styled from 'styled-components';
+import React from "react";
+import SearchNav from "../../../components/Header/SearchHeader";
+import styled from "styled-components";
 
 export default function Search() {
   return (
     <>
-    <SearchNav />
-    <Container>
-    </Container>
+      <SearchNav />
+      <Container></Container>
     </>
-  )
+  );
 }
 
 const Container = styled.div`
@@ -19,4 +18,8 @@ const Container = styled.div`
   align-items: center;
   gap: 24px;
   height: calc(100vh - 108px);
+  overflow-y: scroll;
+  &::-webkit-scrollbar {
+    display: none;
+  }
 `;
