+++ src/pages/Profile/MyProfile.jsx
@@ -1,8 +1,15 @@
 import React from "react";
-import Post1 from "../../components/common/Post/Post1";
+import Post from "../../components/common/Post/Post";
 import MyProfileUp from "../../components/common/Profile/MyProfileUp";
 import ModalNav from "../../components/Header/ModalHeader";
-import { FlexIconImg, GridIconImg, MainWrap, Wrap } from "./MyProfileStyle";
+import {
+  FlexIconImg,
+  GridIconImg,
+  MainWrap,
+  Wrap,
+  Container,
+  PostContainer
+} from "./MyProfileStyle";
 import flexIcon from "../../assets/images/icon-flex-on.svg";
 import gridIcon from "../../assets/images/icon-grid-off.svg";
 
@@ -10,14 +17,19 @@ export default function MyProfile() {
   return (
     <>
       <ModalNav />
-      <MyProfileUp />
-      <MainWrap>
-        <Wrap>
-          <FlexIconImg src={flexIcon} alt="나열방식"></FlexIconImg>
-          <GridIconImg src={gridIcon} alt="그리드방식"></GridIconImg>
-        </Wrap>
-        <Post1 />
-      </MainWrap>
+      <Container>
+        <MyProfileUp />
+        <MainWrap>
+          <Wrap>
+            <FlexIconImg src={flexIcon} alt="나열방식" />
+            <GridIconImg src={gridIcon} alt="그리드방식" />
+          </Wrap>
+          <PostContainer>
+            <Post />
+            <Post />
+          </PostContainer>
+        </MainWrap>
+      </Container>
     </>
   );
 }
